refactor : 변수명 수정 및 불필요한 요소들 정리
