+++ src/main/java/com/coniverse/dangjang/domain/auth/service/OauthInfo/OAuthApiClient.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.auth.service.OAuthInfo;
+package com.coniverse.dangjang.domain.auth.service.OauthInfo;
 
 import com.coniverse.dangjang.domain.auth.dto.OAuthProvider;
 import com.coniverse.dangjang.domain.user.infrastructure.OAuthInfoResponse;
