+++ src/test/java/com/coniverse/dangjang/domain/auth/dto/AuthToken/AuthTokensTest.java
@@ -1,6 +1,6 @@
 package com.coniverse.dangjang.domain.auth.dto.AuthToken;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
  * @author EVE
  * @since 1.0
  */
-public class AuthTokensTest {
+class AuthTokensTest {
 	@Test
 	void AuthToken_객체_생성() {
 		AuthTokens token = new AuthTokens("accessToken", "refreshToken", "B", 1800L);
