+++ src/test/java/com/coniverse/dangjang/domain/auth/service/OAuthInfoServiceTest.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang.domain.user;
+package com.coniverse.dangjang.domain.auth.service;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -6,15 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import com.coniverse.dangjang.domain.auth.dto.request.KakaoLoginParams;
-import com.coniverse.dangjang.domain.auth.service.ProductOAuthInfoService;
 
 @SpringBootTest
-public class OAuthInfoServiceTest {
+class OAuthInfoServiceTest {
 	@Autowired
 	private ProductOAuthInfoService oAuthInfoService;
 
 	@Test
-	public void accessToken으로_사용자정보_가져오기() {
+	void accessToken으로_사용자정보_가져오기() {
 		KakaoLoginParams kakaoLoginParams = new KakaoLoginParams();
 		kakaoLoginParams.setAccessToken("4J-zgwK68lN3RIm8iy1Qv0EGE54mbyOrVc-X1cf1CinJXgAAAYk1SMch");
 		kakaoLoginParams.oAuthProvider();
