+++ src/test/java/com/coniverse/dangjang/domain/auth/service/OauthInfo/NaverApiClientTest.java
@@ -1,6 +1,6 @@
-package com.coniverse.dangjang.domain.auth.service.OAuthInfo;
+package com.coniverse.dangjang.domain.auth.service.OauthInfo;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,12 +11,12 @@ import org.springframework.boot.test.context.SpringBootTest;
  * @since 1.0
  */
 @SpringBootTest
-public class NaverApiClientTest {
+class NaverApiClientTest {
 	@Autowired
 	private NaverApiClient naverApiClient;
 
 	@Test
-	public void 네이버_사용자_정보_가져오기_실패한다() throws Exception {
+	void 네이버_사용자_정보_가져오기_실패한다() throws Exception {
 		assertThrows(Exception.class, () -> {
 			naverApiClient.requestOauthInfo("");
 			;
